iteration 1 but with a ContactTable component imported to AllContacts
