+++ src/components/all-contacts/AllContacts.css
@@ -1,0 +1,3 @@
+h1 {
+    text-align: center;
+}
